Fix method ambiguity in TracedRNG seed!
`Random.seed!(trng::TracedRNG, key)` took an untyped `key`, making it ambiguous with
`Random.seed!(::AbstractRNG, ::Nothing)` (Random stdlib) for the call
`seed!(::TracedRNG, ::Nothing)`. The Aqua ambiguity check flags this on CI (the
conflicting Random method isn't present in every local Random, so it didn't reproduce
locally). The seed is always an integer, so constrain `key::Integer`, which removes the
overlap with `::Nothing`.

+++ src/mcmc/particle_mcmc.jl
@@ -64,7 +64,7 @@ Random.rand(trng::TracedRNG, ::Type{T}) where {T<:Unsigned} = Random.rand(trng.r
 inner_key(rng::Random123.Philox2x) = rng.key
 
 "Reseed and rewind the inner generator. The model-step counter is left untouched."
-function Random.seed!(trng::TracedRNG, key)
+function Random.seed!(trng::TracedRNG, key::Integer)
     Random.seed!(trng.rng, key)
     Random123.set_counter!(trng.rng, 0)
     return trng
